feat: Change service model to batched jobs instead
