+++ .gitignore
@@ -4,3 +4,4 @@ dist
 .vscode
 .env
 .vercel
+absen.log
