+++ batch/tibsen.bat
@@ -1,0 +1,1 @@
+node %dp0/../dist/src/index.js
