--- nodemon.json
@@ -1,16 +1,0 @@
-{
-  "restartable": "rs",
-  "verbose": true,
-  "watch": [
-    "src/**/*.ts"
-  ],
-  "ignore": [
-    ".git",
-    "node_modules/**"
-  ],
-  "env": {
-    "NODE_ENV": "development"
-  },
-  "exec": "node -r ts-node/register ./src/index.ts",
-  "ext": "ts"
-}
