+++ package.json
@@ -4,10 +4,8 @@
   "description": "StuPor presence marking made easy ✌️",
   "main": "dist/index.js",
   "scripts": {
-    "dev": "nodemon",
     "build": "rm -rf dist/ && tsc",
-    "start": "pm2 start dist/index.js --name tibsen",
-    "stop": "pm2 stop tibsen",
+    "start": "node dist/index.js",
     "postuninstall": "npm prune"
   },
   "keywords": [],
@@ -17,12 +15,10 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@types/koa": "^2.11.6",
-    "@types/koa-bodyparser": "^4.3.0",
-    "@types/koa-helmet": "^6.0.1",
-    "@types/koa__router": "^8.0.3",
     "@types/luxon": "^1.25.0",
+    "@types/node-cron": "^2.0.3",
     "@types/tough-cookie": "^4.0.0",
+    "@types/winston": "^2.4.4",
     "@typescript-eslint/eslint-plugin": "^4.1.1",
     "@typescript-eslint/parser": "^4.1.1",
     "eslint": "^7.9.0",
@@ -34,13 +30,11 @@
     "typescript": "^4.0.2"
   },
   "dependencies": {
-    "@koa/router": "^9.4.0",
+    "dotenv": "^8.2.0",
     "got": "^11.6.2",
-    "koa": "^2.13.0",
-    "koa-bodyparser": "^4.3.0",
-    "koa-helmet": "^6.0.0",
     "luxon": "^1.25.0",
-    "pm2": "^4.5.0",
-    "tough-cookie": "^4.0.0"
+    "node-cron": "^2.0.3",
+    "tough-cookie": "^4.0.0",
+    "winston": "^3.3.3"
   }
 }
