+++ src/index.ts
@@ -1,19 +1,49 @@
-import Koa from 'koa';
-import helmet from 'koa-helmet';
-import bodyParser from 'koa-bodyparser';
-import { router } from './router/routes';
-import { responseTimeMiddleware } from './util/middleware';
+import { schedule } from 'node-cron';
+import { DateTime } from 'luxon';
+import { createHTTPClient } from './common/client';
+import { SSOAuthenticator } from './services/sso';
+import { StuPorCrawler } from './services/stupor';
+import { isValidConfig } from './common/config';
+import { logger } from './services/logger';
+import { ParameterError } from './exception/parameter';
+import { TibsenError } from './exception/base';
+import config from './../config.json';
 
-const app = new Koa();
+schedule('1 7-17 * * 1-5', async () => {
+  try {
+    isValidConfig(config);
 
-app.use(bodyParser());
-app.use(helmet());
-app.use(router.routes());
-app.use(router.allowedMethods());
+    logger.write(`[${new Date().toISOString()}] Absen protocol initiated`);
 
-app.use(responseTimeMiddleware);
+    const { schedule, email, password } = config;
+    const { weekday, hour } = DateTime.local().setZone('Asia/Jakarta');
 
-const port = process.env.PORT || 3000;
+    if (schedule[weekday - 1].includes(hour)) { // sunday is zero
+      const httpClient = createHTTPClient();
+      const authenticator = new SSOAuthenticator(httpClient, email, password);
+      const crawler = new StuPorCrawler(authenticator);
 
-app.listen(port);
-console.log(`Listening on port ${port}`);
+      const result = await crawler.markPresence();
+
+      if (!result) {
+        throw new ParameterError('Wrong schedule');
+      }
+
+      logger.write(
+        `[${new Date().toISOString()}] Successfully tibsen for ${email}`,
+      );
+    }
+  } catch (err) {
+    logger.write(
+      `[${new Date().toISOString()}] Tibsen failed. Reason:`,
+    );
+
+    const error = err as TibsenError;
+
+    logger.write(
+      `[${new Date().toISOString()}] ${error.name}: ${error.message}`,
+    );
+  }
+});
+
+logger.write(`[${new Date()}] Tibsen started`);
