--- src/router/routes.ts
@@ -1,8 +1,0 @@
-import Router from '@koa/router';
-import absen from '../api/absen';
-
-const router = new Router();
-
-router.post('/absen', absen);
-
-export { router };
