--- vercel.json
@@ -1,10 +1,0 @@
-{
-  "version": 2,
-  "routes": [
-    {
-      "src": "/absen",
-      "methods": ["POST"],
-      "dest": "/api/absen" 
-    }
-  ]
-}
